pyproject.toml
[project]
name = "autometa"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.129.2",
    "uvicorn[standard]>=0.40.0",
    "python-multipart>=0.0.21",

    # Claude API
    "claude-agent-sdk>=0.1.48",

    # Templating
    "jinja2>=3.1.6",
    "markupsafe>=3.0.3",
    "markdown>=3.7",

    # Utils
    "python-dotenv>=1.2.1",
    "requests>=2.33.0",
    "pyyaml>=6.0.2",

    # Database
    "psycopg2-binary>=2.9.10",

    # S3-compatible object storage
    "boto3>=1.35.0",

    # File format support
    "openpyxl>=3.1.0",
    "mammoth>=1.6.0",
    "pdfplumber>=0.10.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.0",
    "bandit>=1.8.0",
    "pip-audit>=2.7.0",
]

# =============================================================================
# Ruff — linting & formatting
# =============================================================================

[tool.ruff]
target-version = "py311"
line-length = 120
# evals/: experimental evaluation scripts, intentionally loose style
# skills/: self-contained agent skill packages with their own conventions
exclude = ["data/", ".venv/", "evals/", "skills/"]

[tool.ruff.lint]
preview = true
select = ["E", "F", "W", "I"]
ignore = [
    "E501",   # line length — enforce progressively, not retroactively
]

[tool.ruff.lint.per-file-ignores]
"web/app.py" = ["E402"]              # imports after logging setup
"scripts/*" = ["E402"]               # sys.path / dotenv setup before imports
"tests/test_knowledge_sync.py" = ["E402"]  # sys.path setup before imports
"tests/test_uploads.py" = ["E402"]         # env config before imports

# =============================================================================
# Bandit — SAST
# =============================================================================

# =============================================================================
# Pytest
# =============================================================================

[tool.pytest.ini_options]
markers = [
    "integration: tests that require external services or credentials",
]

[tool.bandit]
exclude_dirs = ["tests", "evals", "data", "skills"]
skips = [
    "B310",  # urllib.request.urlopen — used for Notion API calls and research sync (controlled URLs)
    "B704",  # markupsafe.Markup — used intentionally to render agent-generated markdown
]
